Move value state from InputBox to App.
